refactor: money Long 타입 변경
- Quantity 메시지 수정

+++ app/src/main/java/woowacourse/shopping/constants/MockData.kt
@@ -11,7 +11,7 @@ object MockData {
         ProductResponse(
             id = i,
             name = "품목$i",
-            price = (i * 1_000),
+            price = (i * 1_000).toLong(),
             imageUrl = "$IMAGE_BASE_URL${(i - 1) % 5}$IMAGE_URL_SUFFIX",
         )
     }
